Rename internal logger module
